+++ src/ipopt_adapter.cc
@@ -26,14 +26,15 @@ bool IpoptAdapter::get_bounds_info(Index n, double* x_lower, double* x_upper,
                                    Index m, double* g_l, double* g_u)
 {
     auto bounds_x = nlp_->GetBoundsOnOptimizationVariables();
-    for (uint c=0; c<bounds_x.size(); ++c) {
+
+    for (size_t c=0; c<bounds_x.size(); ++c) {
         x_lower[c] = bounds_x.at(c).lower_;
         x_upper[c] = bounds_x.at(c).upper_;
     }
 
     // specific bounds depending on equality and inequality constraints
     auto bounds_g = nlp_->GetBoundsOnConstraints();
-    for (uint c=0; c<bounds_g.size(); ++c) {
+    for (size_t c=0; c<bounds_g.size(); ++c) {
         g_l[c] = bounds_g.at(c).lower_;
         g_u[c] = bounds_g.at(c).upper_;
     }
@@ -52,6 +53,8 @@ bool IpoptAdapter::get_starting_point(Index n, bool init_x, double* x,
     assert(init_lambda == false);
 
     VectorXd x_all = nlp_->GetVariableValues();
+
+    // Map read&write access to pointer x
     Eigen::Map<VectorXd>(&x[0], x_all.rows()) = x_all;
 
     return true;
@@ -66,14 +69,18 @@ bool IpoptAdapter::eval_f(Index n, const double* x, bool new_x, double& obj_valu
 bool IpoptAdapter::eval_grad_f(Index n, const double* x, bool new_x, double* grad_f)
 {
     Eigen::VectorXd grad = nlp_->EvaluateCostFunctionGradient(x);
-    Eigen::Map<Eigen::MatrixXd>(grad_f,n,1) = grad;
+    
+    Eigen::Map<Eigen::MatrixXd>(grad_f, n, 1) = grad;
+    
     return true;
 }
 
 bool IpoptAdapter::eval_g(Index n, const double* x, bool new_x, Index m, double* g)
 {
     VectorXd g_eig = nlp_->EvaluateConstraints(x);
+    
     Eigen::Map<VectorXd>(g,m) = g_eig;
+    
     return true;
 }
 
@@ -82,20 +89,23 @@ bool IpoptAdapter::eval_jac_g(Index n, const double* x, bool new_x,
                               double* values)
 {
     // defines the positions of the nonzero elements of the jacobian
-    if (values == NULL) {
+    if(values == NULL){
+
         auto jac = nlp_->GetJacobianOfConstraints();
-        int nele=0; // nonzero cells in jacobian
-        for (int k=0; k<jac.outerSize(); ++k) {
-        for (Jacobian::InnerIterator it(jac,k); it; ++it) {
-            iRow[nele] = it.row();
-            jCol[nele] = it.col();
-            nele++;
-        }
+        int nele = 0; // nonzero cells in jacobian
+        
+        for(int k=0; k<jac.outerSize(); ++k){
+            for(Jacobian::InnerIterator it(jac,k); it; ++it) {
+                iRow[nele] = it.row();
+                jCol[nele] = it.col();
+                nele++;
+            }
         }
 
         assert(nele == nele_jac); // initial sparsity structure is never allowed to change
     }
-    else {
+    else{
+        
         // only gets used if "jacobian_approximation finite-difference-values" is not set
         nlp_->EvalNonzerosOfJacobian(x, values);
     }
